requirements.txt
pytest>=8.0.0
playwright>=1.40.0
pytest-playwright>=0.4.0
allure-pytest>=2.13.0
pytest-xdist>=3.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0